% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egss_mle.R
\name{egss_mle}
\alias{egss_mle}
\title{Compute the MLE of the EGSS model}
\usage{
egss_mle(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess_egss}{A vector of first guess of the four parameters, from \code{guess_egss()}}
}
\value{
A list of MLEs of four parameters, estimated log-likelihood, and Akaike Information Criteria.
}
\description{
This function computes the Maximum Likelihood Estimates for the Exponential Growth State-Space (EGSS) population model.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
egss_mle(yt = yt1, tt = tt1, fguess_egss = guess_egss(yt = yt1, tt = tt1))
}
