% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egss_predict.R
\name{egss_predict}
\alias{egss_predict}
\title{Predict trajectory for EGSS}
\usage{
egss_predict(yt, tt, parms, CIparms, plot.it = "TRUE")
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{parms}{Parameters estimates of REMLEs from `egss_remle()}
}
\value{
A dataframe of the original tt vector, Prediction (REMLE), and observed vector (real)
}
\description{
UNFINISHED function that provides prediction trajectory of a population under EGSS model. It is still in development.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
parms1 = egss_remle(yt = yt1, tt = tt1, fguess_egss = guess_egss(yt = yt1, tt = tt1))

egss_predict(yt = yt1, tt = tt1, parms = parms1$remles, thetaCI = parms1$CIs.theta)

}
