% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egss_remle.R
\name{egss_remle}
\alias{egss_remle}
\title{Compute the REMLE of the EGSS model}
\usage{
egss_remle(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess}{A vector of first guess of the four parameters, from \code{guess_egss()}}
}
\value{
A list of REMLEs of two maximized parameters (sigmasqr and tausqr) and two estimated parameters (theta and x0), estimated log-likelihood, and Akaike Information Criteria.
}
\description{
This function computes the Restricted Maximum Likelihood Estimates for the Exponential Growth State-Space (EGSS) population model.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
egss_remle(yt = yt1, tt = tt1, fguess = guess_egss(yt = yt1, tt = tt1))
}
