% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_egss.R
\name{guess_egss}
\alias{guess_egss}
\title{Generate a vector of parameter guess for EGSS}
\usage{
guess_egss(yt, tt)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}
}
\value{
A vector of four values, the first guess of the parameters to conduct numerical optimization.
}
\description{
The numerical optimization (Maximum Likelihood Estimate - MLE) for the Exponential Growth State-Space (EGSS) population model requires a first guess of the four parameters:
\itemize{
\item ğœƒ (ğ‘Ÿâˆ’1/2ğ›½^2; or population growth rate, as defined by ğœ‡91 in Dennis \emph{et al.} (1991) and ğœ‡ in Humbert \emph{et al.} (2009)),
\item ğœ^2 (variability of process noise),
\item ğœ^2 (variability of sampling).
\item ğ‘¥_0 (initial population in log-scale)
}

This function provides the first guess of these parameters, roughly computed by provide the vector of log-abundance observations \code{yt} (ğ²) and the vector of observation times \code{tt}
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
guess_egss(yt = yt1, tt = tt1)

}
