% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_ouss.R
\name{guess_ouss}
\alias{guess_ouss}
\title{Generate a vector of parameter guess for OUSS}
\usage{
guess_ouss(yt, tt)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}
}
\value{
A vector of four values, the first guess of the parameters to conduct numerical optimization.
}
\description{
The numerical optimization (Maximum Likelihood Estimate - MLE and Restricted Maximum Likelihood Estimates - REMLE) for the Ornstein-Uhlenbeck State-Space (OUSS) population model requires a first guess of the four parameters:
\itemize{
\item ğœ‡ (in stationary distribution is the expected value of stationary distribution of log-abundance mean; it is different than ğœ‡91 in Dennis et al. (1991) and ğœ‡in Humbert et al. (2009), named here ğœƒ in 1.3.2. EGSS),
\item ğœƒ (rate to approach to stationarity),
\item ğ›½^2 (variability of process noise),
\item ğœ^2 (variability of sampling).
}

This function provides the first guess of these parameters, roughly computed by provide the vector of log-abundance observations \code{yt} (ğ²) and the vector of observation times \code{tt}
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
guess_ouss(yt = yt1, tt = tt1)
}
