% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike_egss_remle.R
\name{negloglike_egss_remle}
\alias{negloglike_egss_remle}
\title{Numerical optimization for REMLE of the EGSS}
\usage{
negloglike_egss_remle(fguess_egss, yt, tt)
}
\arguments{
\item{fguess_egss}{A vector of first guess of the four parameters, from \code{guess_egss()}}

\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}
}
\value{
the value of negative log-Likelihood for the EGSS REMLE
}
\description{
Function needed to compute the Restricted Maximum Likelihood Estimation of parameters within the multivariate log-likelihood for the Exponential Growth State-Space (EGSS) population model, see Humbert \emph{et al.} (2009).
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
negloglike_egss_remle(yt = yt1, tt = tt1, fguess_egss = guess_egss(yt = yt1, tt = tt1))
}
