% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike_ouss_remle.R
\name{negloglike_ouss_remle}
\alias{negloglike_ouss_remle}
\title{Numerical optimization for REMLE of the OUSS}
\usage{
negloglike_ouss_remle(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess}{A vector of first guess of the four parameters, from \code{guess_ouss()} (this function will use only 2:4 parameters)}
}
\value{
the value of negative log-Likelihood for the OUSS REMLE
}
\description{
Function needed to compute the Restricted Maximum Likelihood Estimation of parameters within the multivariate log-likelihood for the stationary Ornstein-Uhlenbeck State-Space (OUSS) population model, see Dennis & Ponciano (2014).
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
negloglike_ouss_remle(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))
}
