% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouss_calc.R
\name{ouss_calc}
\alias{ouss_calc}
\title{Estimate parameters, compute predictions and run parametric bootstrapping}
\usage{
ouss_calc(
  yt,
  tt,
  pmethod = "ML",
  nboot,
  plot.pred = "TRUE",
  plot.bootdists = "TRUE"
)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{pmethod}{Select one of two methods: ML or REML}

\item{nboot}{number of bootstrap}

\item{plot.pred}{Logical argument of plot observed and predicted abundances in the time series}

\item{plot.bootdists}{Logical argument of plot histograms of parametric bootstrapping of the four parameters}
}
\value{
A list of parameter estimates, log-likelihood estimate, AIC, matrix of bootstrapped values, parametric bootstrapped confidence intervals, and predictions (with CIs) of abundance (REMLEs or MLEs) for times with observations (1) and including gaps (2)
}
\description{
This function integrates several functions at once. It estimates the parameters using two methods (Maximum Likelihood "ML" or Restricted Maximum Likelihood "REML"), compute the predictions of abundance from the OUSS model, and run a parametric bootstrapping to estimate confidence intervals. In addition, it can generates figures of the bootstrapped parameter estimation and/or the trajectories (observed vs predicted).
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
ouss_calc(yt = yt1, tt = tt1, pmethod = "REML", nboot = 100, plot.pred = TRUE, plot.bootdists = T)

}
