% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouss_pboot.R
\name{ouss_pboot}
\alias{ouss_pboot}
\title{Parametric bootstrapping for OUSS}
\usage{
ouss_pboot(B, yt, tt, parms, REMLE = "FALSE", plot.it = "FALSE")
}
\arguments{
\item{B}{number of bootstraps}

\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{parms}{Parameters estimates (MLEs or REMLEs) from \code{ouss_remle()} or \code{ouss_mle()}}

\item{REMLE}{Logical argument of use of REMLE}

\item{plot.it}{Logical argument of plot histograms of parametric bootstrapping of the four parameters}
}
\value{
list of bootsrtrapped values. If "plot.it = TRUE", it will return histograms of the four parameters
}
\description{
This function conduct a parametric bootstrapping to estimate confidence intervals (2.5\% to 97.5\%) of parameters and estimates (MLEs and REMLEs) for OUSS model.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
B1 = 10 #recommended for analysis ≥2000
parms1 = ouss_remle(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))

ouss_pboot(B = B1, yt = yt1, tt = tt1, parms = parms1$remle, REMLE=TRUE, plot.it = TRUE)

}
