% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouss_predict.R
\name{ouss_predict}
\alias{ouss_predict}
\title{Predict trajectory for OUSS}
\usage{
ouss_predict(yt, tt, parms, plot.it = "TRUE")
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{parms}{Parameters estimates (MLEs or REMLEs) from \code{ouss_remle()} or \code{ouss_mle()}}

\item{plot.it}{Logical argument of plot observed and predicted abundances in the time series}
}
\value{
If "plot.it = TRUE", it will return a simple figure of the population trend.

A list of:
\enumerate{
\item predicted values for each time point with observed data,
\item predicted values including missing time points (not observed data)
}
}
\description{
This function provides an easy way to predict the trajectory of a population under OUSS model
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
parms1 = ouss_remle(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))

ouss_predict(yt = yt1, tt = tt1, parms = parms1$remles, plot.it = TRUE)

}
