% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouss_remle.R
\name{ouss_remle}
\alias{ouss_remle}
\title{Compute the REMLE of the OUSS model}
\usage{
ouss_remle(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess}{A vector of first guess of the four parameters, from \code{guess_ouss()} (this function will use only 2:4 parameters)}
}
\value{
A list of MLEs of four parameters, estimated log-likelihood, and Akaike Information Criteria.
}
\description{
This function computes the Restricted Maximum Likelihood Estimates for the stationary Ornstein-Uhlenbeck State-Space (OUSS) population model.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
ouss_remle(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))
}
