% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUSS_St_MLE.R
\name{OUSS_St_MLE}
\alias{OUSS_St_MLE}
\title{Compute the MLE of the OUSS model}
\usage{
OUSS_St_MLE(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess}{A vector of first guess of the four parameters, from \code{guess_ouss()}}
}
\value{
list of MLEs of four parameters, estimated log-likelihood, and Akaike Information Criteria.
}
\description{
This function comput the Maximum Likelihood Estimates for the stationary Ornstein-Uhlenbeck State-Space (OUSS) population model.
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
OUSS_St_MLE(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))
}
