% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randmvn.R
\name{randmvn}
\alias{randmvn}
\title{Multivariate normal random number generator}
\usage{
randmvn(n, mu.vec, cov.mat)
}
\arguments{
\item{n}{the number of random samples of a multivariate normal vector}

\item{mu.vec}{mean vector of the multivariate normal distribution to sample from}

\item{cov.mat}{Variance-covariance matrix of the multivariate normal distribution to sample from}
}
\value{
A matrix from a Cholesky decomposition (factorization of a real symmetric positive-definite sqr matriz)
}
\description{
This function generates random numbers from a multivariate normal distribution, required for parametric bootstrapping (prediction and simulation of state space models).
}
\examples{
See `ouss_sim()`

}
