% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_decimal.R
\name{time_to_decimal}
\alias{time_to_decimal}
\title{Convert time (Hour Minute Second) to decimal number.}
\usage{
time_to_decimal(x)
}
\arguments{
\item{x}{A character vector of hour minute second}
}
\value{
A numeric vector of decimal time
}
\description{
This function helps to convert time information (Hour Minute Second) to decimal number.
It was extracted from "Best Practices for Using eBird Data" by Matthew Strimas-Mackey, Wesley M. Hochachka, Viviana Ruiz-Gutierrez, Orin J. Robinson, Eliot T. Miller, Tom Auer, Steve Kelling, Daniel Fink, & Alison Johnston.
This function requires the package \code{lubridate}, included in \code{tidyverse}.
}
\examples{
x <- "5:30:00 AM"
time_to_decimal(x)
}
