% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpm.R
\name{vpm}
\alias{vpm}
\title{Conduct Viable Population Monitoring - a risk-based approach}
\usage{
vpm(yt, tt, ntraj, N.critical, plt)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{ntraj}{Number of trajectories (e.g., 100, 1000) to simulate}

\item{N.critical}{Critical value of the population (recommended = 3/4*mean(observed in the time series))}

\item{plt}{Population Long Term simulations. How long you want to simulate the trajectories?}
}
\value{
A dataframe with the vector of observations times, vector of abundance observations, and two computed probability of extinctions (ntraj below N.critical/ntraj; area under the curve in log-normal distribution below N.critical)
}
\description{
This function provides an easy way to conduct a risk-based viable population monitoring for a time-series data
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
ntraj = 100 #It is better if you simulate over 50000 (the probability stabilizes)
N.critical = 3/4*mean(exp(yt2))
plt = 5

vpm(yt = yt1, tt = tt1, ntraj = ntraj, N.critical = N.critical, plt = plt)
}
