% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negloglike_OU_mle.R
\name{negloglike_OU_mle}
\alias{negloglike_OU_mle}
\title{Numerical optimization for Maximum Likelihood Estimation - OUSS}
\usage{
negloglike_OU_mle(yt, tt, fguess)
}
\arguments{
\item{yt}{A vector of log-abundance observations in the time-series.}

\item{tt}{A vector of observations times.}

\item{fguess}{A vector of first guess of the four parameters, from \code{guess_ouss()}}
}
\value{
value of negative log-Likelihood for the OUSS MLE
}
\description{
The multivariate log-likelihood for the stationary Ornstein-Uhlenbeck State-Space (OUSS) population model is given by:

$$\ln{L}(\mu, \theta, \beta^2, \tau^2) = -\frac{(q+1)}{2}\ln(2\pi)-\frac{1}{2}\ln(|\mathbf{V}|)-\frac{1}{2}(\mathbf{y}-\mathbf{m})'\mathbf{V}^{-1}(\mathbf{y}-\mathbf{m})$$

where $q$ is the number of time-series transitions (thus, $q+1$ reflect the length of the time-series, with the initial population estimation $y_0$ as a realized value of the random variable $Y(0)$),
$\mathbf{V}$ is the variance-covariance matrix (with diagonal computed from $V\link{Y(t_i)} = \tau^2+\frac{\beta^2}{2\theta}$; Eq. 17 in Dennis & Ponciano 2014),
$\mathbf{y}$ is the data values ($y_0$, $y_1$, $y_2$, ..., $y_q$), and
$\mathbf{m}$ is the vector of same $\mu$ in all $q+1$ times ($E\link{Y(t_i)} = \mu]$; Eq. 16 in Dennis & Ponciano 2014).

This function provides the first guess of these parameters, roughly computed by provide the vector of log-abundance observations \code{yt} (𝐲) and the vector of observation times \code{tt}
}
\examples{
yt1 = log(c(18,10,9,14,17,14,5,10,9,5,11,11,4,5,4,8,2,3,9,2,4,7,4,1,2,4,11,11,9,6))
tt1 = c(1966:1995)
negloglike_OU_mle(yt = yt1, tt = tt1, fguess = guess_ouss(yt = yt1, tt = tt1))
}
