% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_decimal.R
\name{time_to_decimal}
\alias{time_to_decimal}
\title{Convert time (Hour Minute Second) to decimal number. This function came from "Best Practices for Using eBird Data" by Matthew Strimas-Mackey, Wesley M. Hochachka, Viviana Ruiz-Gutierrez, Orin J. Robinson, Eliot T. Miller, Tom Auer, Steve Kelling, Daniel Fink, & Alison Johnston}
\usage{
time_to_decimal(x)
}
\arguments{
\item{x}{A character vector of hour minute second}
}
\value{
A numeric vector of decimal time
}
\description{
Convert time (Hour Minute Second) to decimal number. This function came from "Best Practices for Using eBird Data" by Matthew Strimas-Mackey, Wesley M. Hochachka, Viviana Ruiz-Gutierrez, Orin J. Robinson, Eliot T. Miller, Tom Auer, Steve Kelling, Daniel Fink, & Alison Johnston
}
\examples{
x <- "7 6 5"
time_to_decimal(x)
}
